requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.97.0
cachetools==6.2.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.3
cryptography==47.0.0
dataclasses-json==0.6.7
deprecation==2.1.0
distro==1.9.0
docstring_parser==0.18.0
fastapi==0.136.1
frozenlist==1.8.0
fsspec==2026.3.0
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1
google-cloud-core==2.5.1
google-cloud-modelarmor==0.5.0
googleapis-common-protos==1.74.0
greenlet==3.5.0
grpcio==1.80.0
grpcio-status==1.80.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.11
jiter==0.14.0
jsonpatch==1.33
langchain==1.2.15
langchain-anthropic==1.4.1
langchain-classic==1.0.4
langchain-community==0.4.1
langchain-core==1.3.2
langchain-google-community==3.0.5
langchain-openai==1.2.1
langchain-protocol==0.0.12
langchain-text-splitters==1.1.2
langchain-xai==1.2.2
langgraph==1.1.9
langgraph-checkpoint==4.0.3
langgraph-checkpoint-postgres==3.0.5
langgraph-prebuilt==1.0.11
langgraph-sdk==0.3.13
langsmith==0.7.37
markdown-it-py==4.0.0
marshmallow==3.26.2
mdurl==0.1.2
mmh3==5.2.1
multidict==6.7.1
mypy_extensions==1.1.0
numpy==2.4.4
oauthlib==3.3.1
openai==2.32.0
orjson==3.11.8
ormsgpack==1.12.2
postgrest==2.29.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.2
protobuf==6.33.6
psycopg==3.3.3
psycopg-pool==3.3.0
pure_eval==0.2.3
pyasn1==0.6.3
pyasn1_modules==0.4.2
pydantic==2.13.3
pydantic-settings==2.14.0
pydantic_core==2.46.3
pyiceberg==0.11.1
PyJWT==2.12.1
pyparsing==3.3.2
pyroaring==1.1.0
python-dotenv==1.2.2
realtime==2.29.0
regex==2026.4.4
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.3.4
sniffio==1.3.1
SQLAlchemy==2.0.49
starlette==1.0.0
storage3==2.29.0
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.29.0
supabase-auth==2.29.0
supabase-functions==2.29.0
tenacity==9.1.4
tiktoken==0.12.0
tqdm==4.67.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.46.0
websockets==15.0.1
xxhash==3.7.0
yarl==1.23.0
zstandard==0.25.0
